Added a way to show images in WebGL.
Tried to keep this simlar to the existing display buffer handling (i.e. used texture new and drop to create and destroy the texture in WebGL).  Might make sense to have textures stored in Rc+RefCell instances, so they'll be dropped when they're no longer used.

Also added a bunch of panics just in case TypeScript handling breaks.  So now the Rust will at least know that it happens (though it may not respond very gracefully).