Refactored the Tiled file loading.
Now there's a SharedTiledFile instance that is used to access all loading. Multiple files can be loaded at once, so long that they've got different URLs. The core TiledFile object is allocated up-front, and then filled up.

Also replaced the janky `static mut` (used to store the result) with a method on Game. Probably not perfect, but looks way better.