Now jumping off a track and jumping off the ground both "use up" the jump input.
So if a user holds the jump button while trying to snap to a track, they won't immediately jump the frame after snapping to the track. There needs to be a separate press to do that.

This also led to me just simplifying a lot of the code to have simple "input used up" flags.